more elegant prepare script

+++ prepare.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+cp conf/start/zentral/base.json conf/start/zentral/base.json.bak
+
 TLSSECRET=$(cat /dev/urandom | env LC_CTYPE=C tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
 DJANGOSECRET=$(cat /dev/urandom | env LC_CTYPE=C tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
 sed -i "s/API\ SECRET\ \!\!\!\ CHANGE\ THIS\ \!\!\!\ DO\ NOT\ USE\ IN\ PRODUCTION\ \!\!\!/zentral/$TLSSECRET/g" conf/start/zentral/base.json
@@ -17,20 +19,20 @@ echo "Basic auth username:"
 read BAUSERNAME
 echo "Basic auth password:"
 read BAPASSWORD
-rm conf/basic_auth/docker/nginx/extra/zentral.htpasswd
+mv conf/basic_auth/docker/nginx/extra/zentral.htpasswd conf/basic_auth/docker/nginx/extra/zentral.htpasswd.bak
 htpasswd -cb conf/basic_auth/docker/nginx/extra/zentral.htpasswd $BAUSERNAME $BAPASSWORD
 echo "*.htpasswd" >> .gitignore
 
 echo "Enter the full path to your tls certificate:"
 read TLSCERTIFICATE
-rm conf/start/docker/tls/zentral.crt
-mv "$TLSCERTIFICATE" conf/start/docker/tls/zentral.crt
+mv conf/start/docker/tls/zentral.crt conf/start/docker/tls/zentral.crt.bak
+cp "$TLSCERTIFICATE" conf/start/docker/tls/zentral.crt
 echo "*.crt" >> .gitignore
 
 echo "Path to your tls key:"
 read TLSKEY
-rm conf/start/docker/tls/zentral.key
-mv "$TLSKEY" conf/start/docker/tls/zentral.key
+mv conf/start/docker/tls/zentral.key conf/start/docker/tls/zentral.key.bak
+cp "$TLSKEY" conf/start/docker/tls/zentral.key
 echo "*.key" >> .gitignore
 
 sed -i "s/runserver/gunicorn/g" docker-compose-basic-auth.yml
@@ -52,4 +54,8 @@ echo "DB name:"
 read DBNAME
 sed -i "s/POSTGRES_DB=zentral/POSTGRES_DB=$DBNAME/g" conf/start/docker/postgres.env
 
-git rm --cached -r .
+echo "Reset git cache? y or n"
+read ANSWER
+if [[ $ANSWER == 'y' ]]; then
+  git rm --cached -r .
+fi
